[roryhe]-[clean up]-[task-four 个人博客后端接口 文章功能]

+++ go-record/study-base/task-four/models/post.go
@@ -6,7 +6,7 @@ type Post struct {
 	gorm.Model
 	Title    string    `json:"title" gorm:"type:varchar(255)"`
 	Content  string    `json:"content" gorm:"type:text"`
-	UserId   int64     `json:"user_id" gorm:"not null;index"`
+	UserId   uint      `json:"user_id" gorm:"not null;index"`
 	User     User      `json:"user" gorm:"references:ID"`
 	Comments []Comment `json:"comments" gorm:"foreignKey:ID"`
 }
